TFADecoder: extend decoder's cuda() and cpu() methods to handle Variables
The torch.nn.Module().cuda()/cpu() methods only automatically transfer
Parameters between the CPU and GPU, but we can only run inference successfully
when both Variables and Parameters are in the same kind of memory.  As a
convenience, extend the relevant methods in TFADecoder to move the variables
for us.

Testing Done: ran optimization with CUDA!

Signed-off-by: esennesh <elisennesh@gmail.com>