htfa_torch.utils.GaussianTower: a class for handling towers of Gaussians
Sometimes you have a tower of Gaussians: an actual parameter you want to
sample, and the hyperpriors on it that let you be fully Bayesian.  Since this
has a branching factor of two as you add levels, let's make a class to do the
recursion explicitly.

Testing Done: run inference with Gaussian towers in the encoder

Signed-off-by: Eli Sennesh <e.sennesh@northeastern.edu>