dtfa_models.DeepTFAGenerativeHyperparams: add variance to match TFA prior
The original TFA allowed for more variance in factor log-widths and centers
than we seem to be doing, and it seems to cause us to underfit the data on
Sherlock (though we seem to do ok on AffVids).  Try widening the permitted
ranges for template centers, log-widths, and for embeddings.

Signed-off-by: Eli Sennesh <e.sennesh@northeastern.edu>