dtfa_models.DeepTFADecoder: consolidate centers and log-widths into Kx4 model
Consolidate the generative model of centers (Kx3) and log-widths (1x1) into a
combined generative model with an initialized bias.  Include a large standard
deviation on the centers' locations, so as to give the generative model more
room to learn and the variational guide (including embeddings) more room for
individual differences.

Testing Done: able to start mini-Pieman run

Signed-off-by: Eli Sennesh <esennesh@ccis.neu.edu>